+++ src/main/java/com/cap/datareporting/service/SysPermissionService.java
@@ -13,5 +13,5 @@ public interface SysPermissionService {
 
 //    List<Permission> find
 
-    SysPermission findPermissionByid(Integer id);
+    SysPermission findPermissionByid(Long id);
 }
